requirements.txt
# Core dependencies for Worm Arena
# C. elegans simulation with GNN-based neural activity prediction

# Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0

# Data handling
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
networkx>=3.1
seaborn>=0.12.0

# Scientific computing
scipy>=1.10.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

